Add Continuous Deployment via TravisCI + Heroku
Adding in CD to allow us to immediately see changes when a push is made.

This is so that we are able to much more easily make changes and get
them released, opposed to having to manually `git push` / `git pull` in
the correct directories. We also take advantage of TravisCI because of
the deal inside the Github Student Pack where we get unlimited free
private CI projects.

The current URLs are:

- [[http://hacknotts-staging.herokuapp.com]] for the `master` branch
- [[http://hacknotts-prod.herokuapp.com]] for the `prod` branch
  - NOTE: The prod branch is **ONLY** to be used when changes need to
    go live, and have been tested thoroughly.