[PROD] Move from Heroku to AWS-S3
As Heroku is actually pretty slow at serving content, we want to move
away from it, and use something a bit quicker. AWS seems to be the
right one - especially due to MLH's partnership with them.

Additionally, instead of running an EC2 instance, or similar, we will
just run the site from S3 - we luckily only have a static site to
serve, therefore won't need a full fancy server.