Move to the AGPL
Because the site is being distributed over a network, the choice of the
GPL as the license is incorrect, and should really be changed to be
brought in line with this.

However, this may require working with any contributors to the site as
we're technically relicensing the project, despite it still being a
Copyleft license.